\name{get.regpos}
\alias{get.regpos}
\alias{sfsplit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Get Positions of Events Within Regular Expression Match}
\description{
A convenience function that obtains event positions from inside a regular expression match.
}
\usage{
get.regpos(gregx)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gregx}{A regular expression of type \code{gregexpr()}}
}
\author{Christopher Steven Marcum}
\seealso{\code{\link{gregexpr}}}
